# Improve renderViaXMLRPC (webwork2/html2xml) and use it in ProblemSetDetail2.pm and SetMaker.pm