--- a/htdocs/js/apps/ProblemSetDetail2/problemsetdetail2.js
+++ b/htdocs/js/apps/ProblemSetDetail2/problemsetdetail2.js
@@ -282,9 +282,7 @@
 				renderArea.html("<iframe id='psr_render_iframe_" + id +
 					"' src='about:blank' frameBorder='0'></iframe>");
 				var iframe = renderArea.find('#psr_render_iframe_' + id);
-				iframe[0].contentWindow.document.open();
-				iframe[0].contentWindow.document.write(data);
-				iframe[0].contentWindow.document.close();
+				iframe[0].srcdoc = data;
 				iFrameResize({ checkOrigin: false, warningTimeout: 20000, scrolling: true, bodyPadding: 0, bodyBackground: '#f5f5f5' }, iframe[0]);
 				iframe[0].addEventListener('load', function() { iframe[0].contentWindow.document.querySelector('.container-fluid').style.padding = '0px'; });
 				resolve();