--- a/htdocs/js/apps/ProblemSetDetail2/problemsetdetail2.js
+++ b/htdocs/js/apps/ProblemSetDetail2/problemsetdetail2.js
@@ -244,6 +244,7 @@
 			ro.showFooter = "no";
 			ro.displayMode = $('#problem_displaymode').val();
 			ro.extra_header_text = "<style>html{overflow-y:hidden;}body{padding:0;background:#f5f5f5;.container-fluid{padding:0px;}</style>";
+			if (window.location.port) ro.forcePortNumber = window.location.port;
 
 			$.ajax({type:'post',
 				url: basicWebserviceURL,