… htdocs/js/legacy/java_init.js
… htdocs/js/vendor/iframe-resizer/LICENSE
… htdocs/js/vendor/iframe-resizer/README.md
… htdocs/js/vendor/iframe-resizer/js/iframeResizer.contentWindow.map
… htdocs/js/vendor/iframe-resizer/js/iframeResizer.map
… htdocs/package.json
… lib/FormatRenderedProblem.pm
… lib/WeBWorK/ContentGenerator.pm
… lib/WeBWorK/ContentGenerator/Instructor/AchievementEditor.pm
… lib/WeBWorK/ContentGenerator/Instructor/GetLibrarySetProblems.pm
… lib/WeBWorK/ContentGenerator/Instructor/GetTargetSetProblems.pm
… lib/WeBWorK/ContentGenerator/Instructor/ProblemSetDetail2.pm
… lib/WeBWorK/ContentGenerator/Instructor/SetMaker.pm
… lib/WeBWorK/ContentGenerator/Instructor/SetMaker2.pm
… lib/WeBWorK/ContentGenerator/Instructor/SetMakernojs.pm
… lib/WeBWorK/ContentGenerator/Problem.pm
… lib/WeBWorK/ContentGenerator/renderViaXMLRPC.pm
… lib/WeBWorK/Utils/DetermineProblemLangAndDirection.pm
… lib/WeBWorK/Utils/LanguageAndDirection.pm
… lib/WeBWorK/Utils/Tasks.pm
… lib/WebworkClient.pm
… lib/WebworkClient/json_format.pl
… lib/WebworkClient/simple2_format.pl
… lib/WebworkClient/simple_format.pl
… lib/WebworkClient/standard_format.pl
… lib/WebworkClient/sticky_format.pl
… lib/WebworkWebservice.pm
… lib/WebworkWebservice/RenderProblem.pm