The formatRenderedProblem method used by renderViaXMLRPC has been improved to make it do many of the things that Problem.pm and GatewayQuiz.pm do when rendering problems.  The renderProblem method (in RenderProblem.pm) that is also in the chain of that rendering approach has been modified a bit to allow the use of MathQuill answer boxes if enabled for the course.

In ProblemSetDetail2 and in SetMaker, this now loads problems asynchronously via javascript through renderViaXMLRPC.pm instead of loading all problems on the initial page load.  Each problem is loaded successively instead of all at one time, to avoid bogging down the browser.  The point of this change is that many problems currently do not work on the problem set detail page or in the library browser.  This puts each problem into its own iframe and solves those issues.  Most likely all problems will now work.

Previously when problems were loaded via javascript after the page loads the instructorXMLHandler was used.  Now renderViaXMLRPC is used for this as well.

The simple_format.pl needed some minor changes to make this work well.

The json_format.pl has been cleaned up and improved also.  In the case that this format is used the header is set with the appropriate content type.  Related to this is a special case added for the library browser (SetMaker.pm).  If the option `send_pg_flags` is set then the return will also be in json format, but will only contain two keys.  Those are
`html` and `pg_flags`.  The `html` key contains the full html of the format, and the `pg_flags` the `flags` of the pg problem.  This allows the library browser to obtain the insructor comments.  This could be useful in other scenarios as well.

The file lib/WeBWorK/Utils/DetermineProblemLangAndDirection.pm has been renamed to lib/WeBWorK/Utils/LanguageAndDirection.pm and the methods in that file improved to simplify their usage throughout the code.

This also removes the local iframe-resizer javascript and installs a newer version (4.3.1) via npm.

The legacy javascript file htdocs/js/legacy/java_init.js has been deleted and its usage removed from the code.

I also added some improvements to the appearance of the problem set detail 2 page, and eliminated the addition of css style classes via javascript.

This is targeted for the develop branch for now, but it would be nice to get it into the release.