@Alex-Jordan 2021-04-04
Testing this out. So far it works, and I just have cosmetic/layout/design observations in the Library Browser. At least so far, I'm not looking at the code, so I may be off base and citing things that aren't part of this, but I'm seeing them stand out now.

* In the Library Browser, if I reveal the file path of a problem and the file path is even medium length, it butts into the usage statistics and expands the width of the entire block. It's jarring. (a) Can we think of a reason to not simply always show file paths? (b) Could the file path be positioned on the line under the usage statistics, randomize button, etc?
* Also in the Library Browser, I find it jarring when I click to add a problem to a set, and the "(in target set)" text gets inserted between the usage statistics and the icons. Would it be good right-aligned under the icons? (In combination with the previous item, now two things would be on the second line. But less likely to butt into each other.)
* I'm finding it hard to intuit whether a given "Add" button applies to the problem above or below that Add button. It's easy when you are at the top of the list or the problem has been grouped with similar problems so it gets the black border around it and the Add button. But otherwise, viewing the middle of the list, it's unclear. It gets even more muddled if I click something that brings in the results table. I propose styling things in some way that makes it clear which exercise the Add button is tied to. That may mean thinking about the next item:
* I propose changing the styling for the grouped problems. Instead of a border (which might be useful to address the Add button issue) maybe something that indicates an accordion or carousel. Instead of the M button, there could be something that lets you cycle through a carousel.