@drgrice1 2021-04-04
Yeah, so those are all things that need improvement, but those are also all things that were that way before.