@drgrice1 2021-04-04
You must not have looked at the set detail page yet. That has many of these sort of issues as well. I wanted to do a lot more to improve both of these pages, but stopped short before I went to far with this pull request. Perhaps I will try to do that in another pull request yet.