@Alex-Jordan 2021-04-04
What is new is the Preview/Check/ShowCorrectAnswers buttons. Their presence and visual similarity to the Add button makes it less clear that the Add button which is immediately below them is actually paired with the following exercise, not the same exercise those three buttons go with. And then if you use them, the results table appears, also adding a new element of sprawl.