@drgrice1 2021-04-05
I don't think there is quite as much difficulty in determining which problem the add button goes with as you say.  The problem is contained in the gray box, and when the page first loads the add button is at the top of the first problem.  Really, none of those things you are mentioning are changed by this pull request.

In any case, one way to fix this is just to put the border lines around all problems.  However, doing what you suggest for the problems with more is to progressive at this point, and would take a lot more work and is beyond what needs to be done.

Another option is to remove the problem submit buttons (and just make them inert viewable problems instead of active workable).  I had debated on creating a new "even simpler" format that would be used here that does that (doing so with the simple format could be achieved, but is not a very efficient way to handle it).