@drgrice1 2021-04-05
Another simple solution that would work for now (better could be done later) would be to just put a horizontal rule between the problems.