@Alex-Jordan 2021-04-05
The other new thing is that before, the width of a problem is 100%, minus the side navigation panel, but with this branch, something about the iframe and its surrounding div is coming out at a restricted width. And this is what makes the first two items in my list "new". With a wider problem, the butting up doesn't happen except very rarely.

> I don't think there is quite as much difficulty in determining which problem the add button goes with as you say. 

So I really was mixed up which was which, after scrolling down halfway through a page of with 40 OPL problems. I'm not just hypothesizing; it's empirical UX data. I'm sure it's all about the "Add" button for problem 23 falling right under the Preview etc buttons for problem 22. The Preview etc buttons clearly go with 22 because of the background. It just feels more natural (to me at minimum) that the "Add" goes with those buttons. And only my experience that the "Add" goes above a problem gave my brain a counterargument, which a new instructor user doesn't have. Yes, I could go back to the top to reorient myself, and commit to memory what the scheme is, and then try to scroll back down and find the problem I was at before. But why would I do that if I'm in a state of mind where the "Add" button is for 22?

I think all I am saying is that the user experience elements are in a place where we should do something (and I'd be happy to work on this branch too) before merging. I'm not sure how I feel about merging something, even into develop, when we know it needs follow-up work. Even more important if it goes into 2.16, which you mentioned in the opening. I'd like this in 2.16 as well, since I'd like to use more javascript-reliant problems with simple_format.pl.

If you don't beat me to it, the next chunk of time I do things here, I'll put into trying things on this branch. That would be Tuesday or Thursday at the earliest.