@drgrice1 2021-04-05
The width restriction is not new either.  Again, everything you are talking about is the way it was before.

> I'm not just hypothesizing; it's empirical UX data.

Say what now?