@drgrice1 2021-04-05
The point is that none of the issues you are discussing are caused by this pull request.  All of them existed before this pull request.

It is also quite debatable that any of those issues are made worse by this pull request as you claim.  Your empirical UX data set is limited to one data value (you).   That outcome is quite different than my empirical data set with my one data value (me).  Putting that data set together we have two data values with completely different outcomes.  Statistically speaking we can make no conclusion.

Furthermore, the changes that you are requesting are difficult to obtain.  This is already an extensive pull request, and adding these things would make it more so.

The changes in this pull request are more important than the minor visual issues that you are observing, and those can be dealt with later.