@drgrice1 2021-04-05
@Alex-Jordan:  I think that my recent commits have addressed your concerns.  I think that your suggested changes to the "More/Less" implementation would be nice, but are more extensive and can be done later.