@Alex-Jordan 2021-04-05
> The point is that none of the issues you are discussing are caused by this pull request. All of them existed before this pull request.

This is not the case. Probably words can't explain it, and I will make a video to show you the differences. And they are things that make using WeBWorK worse, so they should not be merged yet.