@drgrice1 2021-04-05
This is the case.  You can make a video, and it still won't change that fact.  I see what you are trying to say are differences.  They aren't.  The primary difference is your perception that when buttons have the same style and are anywhere near each other, then they are grouped together.  Even if there are clear indicators that they are not grouped together, like a gray container around the ones above, and the "Add" button not in that below.  My latest commits to this make it painstakingly clear that those buttons are not grouped.