@drgrice1 2021-04-09
> This might be a dumb question, but is it possible to condition the use of an iframe on whether or not the problem loads its own javascript?

That isn't really feasible.  There is no reliable way to determine if a problem uses its own javascript.  Any problem could load an external javascript via the TEXT method, or can even inject javascript directly into the problem without loading an external script.  Furthermore, there are other issues besides javascript.  If you did that the appearance of the problems would be inconsistent.  You would also be adding to the work the server has to do because it would have to preprocess all of the problems again the way it did before.

Have you tried this pull request with the latest improvements?  The loading time is considerably less now.  The first few problems are ready to work almost immediately, and the rest load quite quickly while you are looking at those.  Furthermore, with the last changes, this actually reduces the load on the server from the current setup in many cases.  For example with the problems you mentioned (Calculus single variable - differentiation - quotient rule without trig - show 20), there are 47 problems loaded.  With the current setup, all 47 of those problems are processed by the server.  With this pull request only 20 of them are (although more will be processed on demand).

I did some timing.  I loaded your example (Calculus single variable - differentiation - quotient rule without trig - show 20) on the current WeBWorK-2.16 branch and compared that to this pull request.  The timing was from when the "View Problems" button was pressed (via a submit handler) to the time the the window load event was called (for the WeBWorK-2.16 branch) or to the time the last iframe's load event was called (for this pull request).  I used localStorage to save the start time so the javascript could compute the duration at the end.  On my computer it took on average about 6.5 seconds on the WeBWorK-2.16 branch, and about 16 seconds with this pull request.  With the pull request, I also timed each problem from the submit button press until it was viewable.  The first problem was viewable in less than 2 seconds every time.  The first 8 were viewable in the same 6 - 7 second time frame as the WeBWorK-2.16 branch.

So there is a big difference here that you are failing to observe.  On the WeBWorK-2.16 branch you see nothing until the end of those 6 to 7 seconds.  While with this pull request you have some problems ready to go in less than that, and 8 problems in the same time as the current setup.  Those first problems are fully set up and even workable in that 6 to 7 seconds.

This test was performed with my more than 10 year old computer running both the browser and the server.  This should be considerably less with a more modern computer that is not also running the server.

So if you are still thinking that this is that slow, then I have to ask ... how slow is your computer?